VSpython/pythonfunction/requirements.txt
azure-functions==1.18.0
azure-functions-worker==1.0.0